# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForSEII2526.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Bonos/actionCompraBono": {
      "get": {
        "tags": [
          "Bonos"
        ],
        "operationId": "GetBonosForCompra",
        "parameters": [
          {
            "name": "nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipobocadillo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BonosForCompraDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BonosForCompraDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BonosForCompraDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/CompraBonos/GetCompraBono": {
      "get": {
        "tags": [
          "CompraBonos"
        ],
        "operationId": "GetCompraBono",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BonosDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BonosDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BonosDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/CompraBonos/CrearCompra": {
      "post": {
        "tags": [
          "CompraBonos"
        ],
        "operationId": "CrearCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BonosForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BonosForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BonosForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BonosDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BonosDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BonosDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationUser": {
        "required": [
          "apellido1",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "minLength": 1,
            "type": "string"
          },
          "apellido1": {
            "minLength": 1,
            "type": "string"
          },
          "apellido2": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BonosDetailDTO": {
        "required": [
          "apellido1",
          "apellido2",
          "metodoPago",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "apellido1": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "apellido2": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "metodoPago": {
            "$ref": "#/components/schemas/EnumMetodosPago"
          },
          "fechaCompra": {
            "type": "string",
            "format": "date-time"
          },
          "precioTotal": {
            "type": "number",
            "format": "double"
          },
          "applicationUser": {
            "$ref": "#/components/schemas/ApplicationUser"
          },
          "itemsBono": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BonosItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BonosForCompraDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "precio": {
            "minimum": 1,
            "type": "number",
            "format": "currency"
          },
          "numeroBocadillos": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "tipoBocadillo": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BonosForCreateDTO": {
        "required": [
          "apellido1",
          "apellido2",
          "metodoPago",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "apellido1": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "apellido2": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "metodoPago": {
            "$ref": "#/components/schemas/EnumMetodosPago"
          },
          "itemsCompraBono": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BonosItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BonosItemDTO": {
        "required": [
          "cantidad"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "precioPorBono": {
            "minimum": 1,
            "type": "number",
            "format": "currency"
          },
          "numeroBocadillos": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "tipoBocadillo": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "minimum": 1,
            "type": "integer",
            "format": "currency"
          }
        },
        "additionalProperties": false
      },
...
      "EnumMetodosPago": {
        "enum": [
          "Tarjeta",
          "PayPal",
          "GPay"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}